third_party/WebKit/LayoutTests/external/wpt/infrastructure/browsers/firefox/prefs-expected.txt

This is a testharness.js-based test.
FAIL Ensure that setting gecko prefs works Uncaught Error: assert_equals: expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
